+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.0.7] - 2015-01-16
 ### Added
 - Dot product, using scikits.cuda
+- Vector addition, using cublas
 
 ## [0.0.6] - 2014-12-15
 ### Bugfix